% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{annotate_potato_l}
\alias{annotate_potato_l}
\title{Geom for hoxo_m}
\usage{
annotate_potato_l(x, y, size = 5, interpolate = FALSE, ...)
}
\description{
This is not about Potato L
}
\examples{
library(ggplot2)

ggplot(data.frame(x = 1:10, y = 1:10), aes(x, y)) +
  annotate_potato_l(5,5, size = 3) +
  coord_equal()

}

